refactor(plot): delegate stacked_bar()in Bokeh to BokehStackedBar class
